Rename processlist.component.ts to process-list.component.ts Rename processlist.component.md to process-list.component.md
